Support for walking AST with SELECT statement ONLY
We now have support for dealing with SELECT
statements as specified in the ICU grammar, e.g.

In [1]: a =
    g.parse('''{WHO, select, male {He} female {She} other {They}}''')

In [2]: i = ICUNodeVisitor({'WHO': 'male'})

In [3]: i.visit(a)
Out[3]: u'He'

In [4]: i = ICUNodeVisitor({'WHO': 'female'})

In [5]: i.visit(a)
Out[5]: u'She'

In [6]: i = ICUNodeVisitor({'WHO': 'other'})

In [7]: i.visit(a)
Out[7]: u'They'

Currently the 'other' statement is not an `else` branch but works as a
required keyword, although this needs to be changed to come into
line with the ICU messageformat.

Other element_format types (plural and selectordinal) are not
implemented yet.