Add support for curly-braces string replacement
Here we add support for {} based replacements, as is required by the
message_format_element node within the grammar, when specified without
a element_format. This is a familiar type of text-replacement for
strings in Python. Example:

In [1]: a = g.parse('''{NAME} welcomes you.''')

In [2]: i = ICUNodeVisitor({'NAME': 'Siame'})

In [3]: i.visit(a)
Out[3]: u'Siame welcomes you.'

Also within this commit we have changed the functionality of the `other`
keyword, now causing it to act as the else branch of the SELECT
statement, rather than before when it was a required keyword within
the ICUNodeVisitor initialisation dict.
Example:

In [1]: a = g.parse('''{WHO, select, male {He} female {She} other {They}}''')

In [2]: i = ICUNodeVisitor({})

In [3]: i.visit(a)
Out[3]: u'They'