Support for Plural statements, without offsets
Plurality support for the ICUNodeVisitor! This includes replacing
octothorpe with the value passed in, as well as =N for specific
messages. However offsets have not yet been implemented.

Also, I've made a start to the CLDR/ICU cardinality rules, so that we
can correctly pick the level of plurality(?) required for the given
number and locale.

The walker has stood up well to manual testing, however its
implementation depending on datatypes of incoming data makes me doubt
its robustness, so I'll need to rewrite this to be a bit better in
its decision on whether a plural-type change or selection-type change
is to be made.