Match type of replacement and act accordingly
By capturing the type of replacement required for a given key, e.g.
plural or select, we can better handle how the replacement for that
key should take place.

This solves the problem that occurred when passing in strings of
numbers as a plural parameter in the initialisation dict, as it can now
detect and properly match against exact matches.