Add tests for ICUNodeVisitor and locales
This commit adds basic tests for each of the types of replacement
supported by the ICUNodeVisitor: select, plural and direct
replacements.

Also basic tests for the locale defining functions.

This commit also fixes a bug found when no offset is passed in to a
plural statement causing the walker to error out.