Allow decimals to be used as plural arguments
We now support decimals to be used both as arguments passed in for the
ICUNodeVisitor's initialisation dict, as well as for =N replacements.

Support has also been added within locales.py to retrieve the
different parts of a number, that can change the plurality required
for that word.

Also in this commit we add some basic tests for the walker, as well as
for the individual locale functions.